Bug fix when the term is in the first or last node
This is a bug when finding the group on a following level and finding it
in the first part of the non-first part of the top level, or on the last
part of the non-last part of the top level. It was caused because the
end of the interval defined on the previous level was forgotten about in
the search.